refactored api login, created sample test
